chore: change new helper files structure
