+++ apis/supernova/src/handler/__tests__/hypercube-handler.test.js
@@ -1,4 +1,4 @@
-import * as hcHelper from '../utils/hypercube-helper';
+import * as hcHelper from '../utils/hypercube-helper/hypercube-utils';
 import HyperCubeHandler from '../hypercube-handler';
 
 describe('HyperCube Handlers', () => {
