+++ apis/supernova/src/handler/utils/field-helper/get-data-geo-field.js
@@ -1,4 +1,4 @@
-import { isDateField, isGeoField } from '../utils';
+import { isDateField, isGeoField } from './field-utils';
 
 const getDataGeoField = (field) => {
   const item = field;
