+++ apis/supernova/src/handler/utils/field-helper/get-derived-fields.js
@@ -1,4 +1,4 @@
-import { trimAutoCalendarName } from '../utils';
+import { trimAutoCalendarName } from './field-utils';
 
 const getDerivedFields = (field) => {
   const derivedFields = [];
