--- apis/supernova/src/handler/utils/field-helper/index.js
@@ -1,5 +1,0 @@
-export { default as getDataGeoField } from './get-data-geo-field';
-export { default as getDerivedFields } from './get-derived-fields';
-export { default as expandFieldsWithDerivedData } from './expand-field-derived-data';
-export { default as getAutoSortLibraryDimension } from './get-sorted-library-field';
-export { default as findFieldInExpandedList } from './find-field-in-expandedList';
