+++ apis/supernova/src/handler/utils/hypercube-helper/add-main-dimension.js
@@ -1,0 +1,12 @@
+import { insertMainDimension } from './hypercube-utils';
+
+export default function addMainDimension(self, dimension, index) {
+  const dimensions = self.getDimensions();
+  const idx = index ?? dimensions.length;
+
+  if (dimensions.length < self.maxDimensions()) {
+    return insertMainDimension(self, dimension, dimensions, idx);
+  }
+
+  return Promise.resolve();
+}
