+++ apis/supernova/src/handler/utils/hypercube-helper/add-main-measure.js
@@ -1,0 +1,12 @@
+import { insertMainMeasure } from './hypercube-utils';
+
+export default function addMainMeasure(self, measure, index) {
+  const measures = self.getMeasures();
+  const idx = index ?? measures.length;
+
+  if (measures.length < self.maxMeasures()) {
+    insertMainMeasure(measure, measures, idx);
+  }
+
+  return measure;
+}
